MPS: [WIP] Add support for E-ACSL invariant checking
This commit adds a makefile `Makefile_EACSL` which can be used
to use the E-ACSL plugin of Frama-C to convert the ACSL invariant
annotations in the MPS header files into runtime assertions.

Some limitations of E-ACSL need to be manually overcome,
and the code won't compile with the upstream version of E-ACSL
at the moment. Text @hanno-arm for more information until
E-ACSL testing is properly integrated into the test infrastructure.