Writer: Allow partial commit of output buffers
The previous usage-model of the writer is the following:
The user requests a buffer of a given size via mbedtls_writer_get(),
writes data to it, and commits the entire buffer via
mbedtls_writer_commit().

This model assumes that the user knows how much data should be
written. However, currently there are some writing functions in
Mbed TLS for which we don't know in advance how much data they
will write, and instead they take the size of an output buffer
and return the amount of data they have actually written (if
enough space was available).
On the highest level, one class of such functions are the current
handshake message writing functions, which assume a raw buffer
and write a handshake message to it, only noticing in the end
how much space was actually needed.

To allow the use of such functions with the writer objects,
this commit adds a function

mbedtls_writer_commit_partial()

which allows to commit only parts of the output buffers
that have been obtained since the last commit.

Specifically, mbedtls_writer_commit_partial() accepts an
additional argument `omit` which indicates how many trailing
bytes should be left out from the committing.

The reason this function takes the number of bytes to omit
as opposed to the number of bytes to commit is the following:
If it were to take the number of bytes to commit, firstly
one would need to clarify if these concern the last fetch
only or the entire range fetched since the last commit).
In the former case, the implementation would need to track
the offset of the last fetch, which it currently doesn't.
In the latter case, the user would need to keep track of
adding the sizes of the chunks obtained from mbedtls_writer_get(),
which contradicts the purpose of the writer of *hiding*
such arithmetic from the user.

The support for partial commits requires a modification
in the way mbedtls_writer_reclaim() works: Previously,
when the writer uses a queue to serve the user's request,
the overlapping part of queue and output buffer would be
copied from the queue to the output buffer only when
reclaiming. With partial commits, this does no longer work:
If a user makes a large write request which is served from
the queue, but then commits only a small part which could
in fact have been served from the output buffer, then the
copying of the data from the queue to the output buffer
must happen when calling mbedtls_writer_commit_partial()
since afterwards there is no trace anymore of the writer
using the queue, and the current implementation of
mbedtls_writer_reclaim() wouldn't perform any copying
because it looked as if the user has directly written to
the output buffer.

Instead, the new implementation of the writer will copy
data immediately from the queue to the output buffer once
mbedtls_writer_commit() or mbedtls_writer_partial_commit()
are called, while mbedtls_writer_reclaim() can assume that
this has already been done, and can omit any copying.