Writer: Add random test with maximum chunk size larger than queue
The random test should exercise the behavior of the writer in
the situation whether the user requests more space than what's
left in the current output buffer, but where the remaining space
in the output buffer is still larger than what's available in the
queue. In this case, the writer must not serve from the queue.
The previous random test did not test this because the maximum request
size was set to 40 bytes, with a queue size of 50 bytes.