Writer: Allow unspecified total message length in extended writer
Previously, the extended writer needed to be provided the total
length of the message to be written at initialization time, and
it was checked in mbedtls_writer_done() whether the entire message
had indeed been written.

To support the writing of handshake messages the total size of which
is not known in advance, this commit allows the length to be specified
as -1u, in which case mbedtls_writer_done() will omit the afore-
mentioned check. The rest of the code can stay as is.