Layer 2: Remove EPOCH_VALID flag
The EPOCH_VALID flag was initially introduced to allow adding epochs
to be registered with Layer 2 before they are usable for reading
or writing without having them removed by the epoch garbage collection
immediately.

By now, the epoch garbage collection only removes epochs which are
neither used for reading and writing *and* are older than every
epoch which is used for at least reading or writing. In particular,
future epochs which are not yet in use are not removed, and hence
there is no need for the EPOCH_VALID flag anymore.