MPS Layer 2: Fix replay protection
What was meant to be smart was actually undefined behavior... even
when working with unsigned integer types, shifting beyond the width
of the type is undefined. In fact, some architectures reduce the
shift offset modulo the width of the type,  which in case of the
replay protection lead to functionally wrong code.